[android] fix naming for re-open tests
